encoder pt

+++ fairseq/criterions/label_smoothed_cross_entropy.py
@@ -97,7 +97,7 @@ class LabelSmoothedCrossEntropyCriterion(FairseqCriterion):
 
     def get_lprobs_and_target(self, model, net_output, sample):
         lprobs = model.get_normalized_probs(net_output, log_probs=True)
-        target = model.get_targets(sample, net_output)
+        target = model.get_targets(sample, net_output).long()
         if self.ignore_prefix_size > 0:
             # lprobs: B x T x C
             lprobs = lprobs[:, self.ignore_prefix_size :, :].contiguous()
